+++ grpc-tunnel-client-demo/src/main/java/top/redjujubetree/demo/client/handler/AckRequestMessageHandler.java
@@ -1,6 +1,8 @@
 package top.redjujubetree.demo.client.handler;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import top.redjujubetree.grpc.tunnel.client.GrpcTunnelClientService;
 import top.redjujubetree.grpc.tunnel.client.handler.AbstractServerRequestMessageHandler;
 import top.redjujubetree.grpc.tunnel.proto.RequestPayload;
 import top.redjujubetree.grpc.tunnel.proto.ResponsePayload;
@@ -8,6 +10,8 @@ import top.redjujubetree.grpc.tunnel.proto.ResponsePayload;
 @Component
 public class AckRequestMessageHandler extends AbstractServerRequestMessageHandler {
 
+	@Autowired
+	private GrpcTunnelClientService grpcTunnelClientService;
 	@Override
 	protected boolean supportRequestType(String request) {
 		return "notifyWithAck".equals(request);
