+++ grpc-tunnel-common/src/main/java/top/redjujubetree/grpc/tunnel/constant/ClientRequestTypes.java
@@ -3,5 +3,5 @@ package top.redjujubetree.grpc.tunnel.constant;
 public class ClientRequestTypes {
 	public static final String CONNECT = "_CONNECT";
 	public static final String DISCONNECT = "_DISCONNECT";
-	public static final String HEARTBEAT = "HEARTBEAT";
+	public static final String HEARTBEAT = "_HEARTBEAT";
 }
