+++ grpc-tunnel-server-spring-boot-starter/src/main/java/top/redjujubetree/grpc/tunnel/server/autoconfigure/GrpcTunnelServerAutoConfiguration.java
@@ -12,6 +12,7 @@ import org.springframework.context.annotation.Configuration;
 import top.redjujubetree.grpc.tunnel.handler.MessageHandler;
 import top.redjujubetree.grpc.tunnel.server.GrpcTunnelServerService;
 import top.redjujubetree.grpc.tunnel.server.config.GrpcTunnelServerProperties;
+import top.redjujubetree.grpc.tunnel.server.connection.ConnectionManager;
 import top.redjujubetree.grpc.tunnel.server.filter.BasicClientRegistrationFilter;
 import top.redjujubetree.grpc.tunnel.server.filter.ClientRegisterFilter;
 import top.redjujubetree.grpc.tunnel.server.handler.*;
@@ -24,7 +25,7 @@ import java.util.List;
 @ConditionalOnClass(GrpcServerAutoConfiguration.class)
 @AutoConfigureAfter(GrpcServerAutoConfiguration.class)
 @ConditionalOnProperty(prefix = "grpc.tunnel.server", name = "enabled", havingValue = "true", matchIfMissing = true)
-@ComponentScan(basePackages = "com.redjujubetree.grpc.tunnel.server")
+@ComponentScan(basePackages = "top.redjujubetree.grpc.tunnel.server")
 public class GrpcTunnelServerAutoConfiguration {
 
     @Bean
@@ -54,15 +55,22 @@ public class GrpcTunnelServerAutoConfiguration {
     public GrpcTunnelServerService grpcTunnelServerService(
             GrpcTunnelServerProperties properties,
             List<ClientRegisterFilter> clientRegisterFilters,
-            List<ClientConnectionCloseListener> clientConnectionCloseListeners,
+            ConnectionManager connectionManager,
             List<MessageHandler> messageHandlers,
             HeartbeatHandler heartbeatHandler) {
         return new GrpcTunnelServerService(
                 properties,
                 clientRegisterFilters,
-                clientConnectionCloseListeners,
+                connectionManager,
                 messageHandlers,
                 heartbeatHandler
         );
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    public ConnectionManager connectionManager(
+            List<ClientConnectionCloseListener> clientConnectionCloseListeners) {
+        return new ConnectionManager(clientConnectionCloseListeners);
     }
 }
